requirements.txt
python-dotenv
mediapipe
opencv-python
vidgear
yt-dlp # vidgearの依存関係で必要
slackweb